m2_building_blocks/requirements.txt
torch>=1.11.0
black==22.3.0  # https://stackoverflow.com/questions/71673404/importerror-cannot-import-name-unicodefun-from-click
hydra-core

numpy 
h5py
matplotlib
bokeh